requirements.txt
google-cloud-storage
requests
pytest
pytest-mock
pandas pyarrow